+++ lib/CustomButtons/CustomRadioButton.dart
@@ -29,8 +29,14 @@ class CustomRadioButton<T> extends StatefulWidget {
     this.absoluteZeroSpacing = false,
     this.margin,
     this.wrapAlignment = WrapAlignment.start,
+    this.disabledValues = const [],
+    this.disabledColor,
   })  : assert(buttonLables.length == buttonValues.length,
             "Button values list and button lables list should have same number of eliments "),
+        // assert(
+        //     buttonValues.toSet().intersection(disabledValues.toSet()) ==
+        //         disabledValues.toSet(),
+        //     "Disabled values should be present in button values"),
         assert(buttonValues.toSet().length == buttonValues.length,
             "Multiple buttons with same value cannot exist") {
     if (absoluteZeroSpacing) {
@@ -76,6 +82,9 @@ class CustomRadioButton<T> extends StatefulWidget {
 
   final List<String> buttonLables;
 
+  ///List of disabled values
+  final List<T> disabledValues;
+
   ///Styling class for label
   final ButtonTextStyle buttonTextStyle;
 
@@ -86,6 +95,10 @@ class CustomRadioButton<T> extends StatefulWidget {
 
   ///Selected Color of button
   final Color selectedColor;
+
+  ///Disabled Color of button
+  ///If not provided will use [unSelectedColor]
+  final Color? disabledColor;
 
   ///Unselected Color of the button border
   final Color? unSelectedBorderColor;
@@ -108,19 +121,20 @@ class CustomRadioButton<T> extends StatefulWidget {
   final double elevation;
 
   /// Radius for non-shape radio button
+  @Deprecated('Use shapeRadius instead')
   final double radius;
 
   /// Radius for shape radio button
   final double shapeRadius;
 
-  _CustomRadioButtonState<T> createState() => _CustomRadioButtonState<T>();
+  CustomRadioButtonState<T> createState() => CustomRadioButtonState<T>();
 }
 
-class _CustomRadioButtonState<T> extends State<CustomRadioButton<T>> {
-  String? _currentSelectedLabel;
+class CustomRadioButtonState<T> extends State<CustomRadioButton<T>> {
+  T? _currentSelectedValue;
 
-  Color borderColor(index) =>
-      (_currentSelectedLabel == widget.buttonLables[index]
+  Color _borderColor(T e) =>
+      (_currentSelectedValue == e
           ? widget.selectedBorderColor
           : widget.unSelectedBorderColor) ??
       Theme.of(context).primaryColor;
@@ -130,67 +144,71 @@ class _CustomRadioButtonState<T> extends State<CustomRadioButton<T>> {
     super.initState();
     if (widget.defaultSelected != null) {
       if (widget.buttonValues.contains(widget.defaultSelected)) {
-        int index = widget.buttonValues.indexOf(widget.defaultSelected!);
-        _currentSelectedLabel = widget.buttonLables[index];
+        _currentSelectedValue = widget.defaultSelected;
       } else
         throw Exception("Default Value not found in button value list");
     }
   }
 
+  List<T> get buttonValues => widget.buttonValues;
+
+  List<String> get buttonLables => widget.buttonLables;
+
+  List<T> get disabledValues => widget.disabledValues;
+
+  /// This function will select the button and update the state
+  /// THis can be access from outside to change the selected value programatically
+  /// Please note that this will all call the [radioButtonValue] callback
+  void selectButton(T selectedValue) {
+    widget.radioButtonValue(selectedValue);
+    if (mounted) setState(() {});
+    _currentSelectedValue = selectedValue;
+  }
+
   List<Widget> _buildButtonsColumn() {
-    return widget.buttonValues.map((e) {
-      int index = widget.buttonValues.indexOf(e);
+    return buttonValues.map((e) {
+      int index = buttonValues.indexOf(e);
+      bool disabled = disabledValues.contains(e);
+      bool isSelected = _currentSelectedValue == e;
       return Padding(
         padding: EdgeInsets.all(widget.padding),
         child: Card(
           margin: widget.margin ??
               EdgeInsets.all(widget.absoluteZeroSpacing ? 0 : 4),
-          color: _currentSelectedLabel == widget.buttonLables[index]
-              ? widget.selectedColor
-              : widget.unSelectedColor,
+          color: disabled
+              ? widget.disabledColor ?? widget.unSelectedColor
+              : _currentSelectedValue == e
+                  ? widget.selectedColor
+                  : widget.unSelectedColor,
           elevation: widget.elevation,
           shape: widget.enableShape
-              ? widget.customShape == null
-                  ? RoundedRectangleBorder(
-                      borderRadius:
-                          BorderRadius.all(Radius.circular(widget.shapeRadius)),
-                    )
-                  : widget.customShape
+              ? widget.customShape ??
+                  RoundedRectangleBorder(
+                    borderRadius:
+                        BorderRadius.all(Radius.circular(widget.shapeRadius)),
+                  )
               : null,
           child: Container(
             height: widget.height,
             child: MaterialButton(
               shape: widget.enableShape
-                  ? widget.customShape == null
-                      ? OutlineInputBorder(
-                          borderSide:
-                              BorderSide(color: borderColor(index), width: 1),
-                          borderRadius:
-                              BorderRadius.all(Radius.circular(widget.radius)),
-                        )
-                      : widget.customShape
+                  ? widget.customShape ??
+                      OutlineInputBorder(
+                        borderSide:
+                            BorderSide(color: _borderColor(e), width: 1),
+                        borderRadius: BorderRadius.all(
+                            Radius.circular(widget.shapeRadius)),
+                      )
                   : OutlineInputBorder(
-                      borderSide:
-                          BorderSide(color: borderColor(index), width: 1),
+                      borderSide: BorderSide(color: _borderColor(e), width: 1),
                       borderRadius: BorderRadius.zero,
                     ),
-              onPressed: () {
-                widget.radioButtonValue(e);
-                setState(() {
-                  _currentSelectedLabel = widget.buttonLables[index];
-                });
-              },
-              child: Text(
-                widget.buttonLables[index],
-                textAlign: TextAlign.center,
-                overflow: TextOverflow.ellipsis,
-                maxLines: 1,
-                style: widget.buttonTextStyle.textStyle.copyWith(
-                  color: _currentSelectedLabel == widget.buttonLables[index]
-                      ? widget.buttonTextStyle.selectedColor
-                      : widget.buttonTextStyle.unSelectedColor,
-                ),
-              ),
+              onPressed: disabled ? null : () => selectButton(e),
+              child: Text(widget.buttonLables[index],
+                  textAlign: TextAlign.center,
+                  overflow: TextOverflow.ellipsis,
+                  maxLines: 1,
+                  style: textStyle(isSelected, disabled)),
             ),
           ),
         ),
@@ -199,22 +217,25 @@ class _CustomRadioButtonState<T> extends State<CustomRadioButton<T>> {
   }
 
   List<Widget> _buildButtonsRow() {
-    return widget.buttonValues.map((e) {
-      int index = widget.buttonValues.indexOf(e);
+    return buttonValues.map((e) {
+      int index = buttonValues.indexOf(e);
+      bool disabled = disabledValues.contains(e);
+      bool isSelected = _currentSelectedValue == e;
       return Card(
         margin:
             widget.margin ?? EdgeInsets.all(widget.absoluteZeroSpacing ? 0 : 4),
-        color: _currentSelectedLabel == widget.buttonLables[index]
-            ? widget.selectedColor
-            : widget.unSelectedColor,
+        color: disabled
+            ? widget.disabledColor ?? widget.unSelectedColor
+            : _currentSelectedValue == e
+                ? widget.selectedColor
+                : widget.unSelectedColor,
         elevation: widget.elevation,
         shape: widget.enableShape
-            ? widget.customShape == null
-                ? RoundedRectangleBorder(
-                    borderRadius:
-                        BorderRadius.all(Radius.circular(widget.shapeRadius)),
-                  )
-                : widget.customShape
+            ? widget.customShape ??
+                RoundedRectangleBorder(
+                  borderRadius:
+                      BorderRadius.all(Radius.circular(widget.shapeRadius)),
+                )
             : null,
         child: Container(
           height: widget.height,
@@ -222,39 +243,43 @@ class _CustomRadioButtonState<T> extends State<CustomRadioButton<T>> {
           constraints: widget.autoWidth ? null : BoxConstraints(maxWidth: 250),
           child: MaterialButton(
             shape: widget.enableShape
-                ? widget.customShape == null
-                    ? OutlineInputBorder(
-                        borderSide:
-                            BorderSide(color: borderColor(index), width: 1),
-                        borderRadius:
-                            BorderRadius.all(Radius.circular(widget.radius)),
-                      )
-                    : widget.customShape
+                ? widget.customShape ??
+                    OutlineInputBorder(
+                      borderSide: BorderSide(color: _borderColor(e), width: 1),
+                      borderRadius:
+                          BorderRadius.all(Radius.circular(widget.shapeRadius)),
+                    )
                 : OutlineInputBorder(
-                    borderSide: BorderSide(color: borderColor(index), width: 1),
+                    borderSide: BorderSide(color: _borderColor(e), width: 1),
                     borderRadius: BorderRadius.zero,
                   ),
-            onPressed: () {
-              widget.radioButtonValue(e);
-              setState(() {
-                _currentSelectedLabel = widget.buttonLables[index];
-              });
-            },
+            onPressed: disabled ? null : () => selectButton(e),
             child: Text(
               widget.buttonLables[index],
               textAlign: TextAlign.left,
               overflow: TextOverflow.ellipsis,
               maxLines: 1,
-              style: widget.buttonTextStyle.textStyle.copyWith(
-                color: _currentSelectedLabel == widget.buttonLables[index]
-                    ? widget.buttonTextStyle.selectedColor
-                    : widget.buttonTextStyle.unSelectedColor,
-              ),
+              style: textStyle(isSelected, disabled),
             ),
           ),
         ),
       );
     }).toList();
+  }
+
+  TextStyle textStyle(bool isSelected, bool disabled) {
+    if (isSelected) {
+      return widget.buttonTextStyle.selectedTextStyle.copyWith(
+        color: disabled
+            ? widget.buttonTextStyle.disabledColor
+            : widget.buttonTextStyle.selectedColor,
+      );
+    }
+    return widget.buttonTextStyle.textStyle.copyWith(
+      color: disabled
+          ? widget.buttonTextStyle.disabledColor
+          : widget.buttonTextStyle.unSelectedColor,
+    );
   }
 
   @override